+++ guillotina/contrib/dbusers/services/groups.py
@@ -53,9 +53,9 @@ class ListGroups(ListGroupsOrUsersService):
         return {
             "@name": obj.get("@name"),
             "id": obj.get("id"),
-            "title": obj.get("group_name"),
-            "users": obj.get("group_users") or [],
-            "roles": obj.get("group_user_roles") or [],
+            "title": obj.get("name"),
+            "users": obj.get("users") or [],
+            "roles": obj.get("user_roles") or [],
         }