+++ CHANGELOG.rst
@@ -7,6 +7,10 @@ CHANGELOG
 - Adding POST @groups endpoint in dbusers
 - Upgrading postgres version when running test in conftest
   [nilbacardit26]
+- Updated role definitions in permissions.py to include default values for 'guillotina.Member' and 'guillotina.Manager' to ensure the correct parameters are passed to the role.
+- Modified ListGroups and ListUsers services to include '@id' and 'username' fields in the response.
+- Added validation for 'groupname' in CreateGroup service to ensure it is provided and valid.
+  [rboixaderg]
 
 
 7.0.5 (2025-04-03)