+++ CHANGELOG.rst
@@ -9,7 +9,7 @@ CHANGELOG
   [nilbacardit26]
 - Updated role definitions in permissions.py to include default values for 'guillotina.Member' and 'guillotina.Manager' to ensure the correct parameters are passed to the role.
 - Modified ListGroups and ListUsers services to include '@id' and 'username' fields in the response.
-- Added validation for 'groupname' in CreateGroup service to ensure it is provided and valid.
+- Use the group name as the ID in groups if it is provided.
   [rboixaderg]