+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ CHANGELOG
 - Updated role definitions in permissions.py to include default values for 'guillotina.Member' and 'guillotina.Manager' to ensure the correct parameters are passed to the role.
 - Modified ListGroups and ListUsers services to include '@id' and 'username' fields in the response.
 - Use the group name as the ID in groups if it is provided.
+- Updated @users and @groups endpoints to adapt them to volto/plone api
+- Fixed updating users in groups in custom and standard API
   [rboixaderg]